% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfy.R
\name{cfy}
\alias{cfy}
\title{Year from Congress}
\usage{
cfy(year)
}
\arguments{
\item{year}{the numerical value of the year for which you would like to know}
}
\value{
a numeric value or vector of values of the Congress or Congresses which correspond the years
}
\description{
This function takes a year and returns the numerical value of the session of Congress during that year.
}
