% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_to_year.R
\name{congress_to_year}
\alias{congress_to_year}
\title{Congress to Year}
\usage{
congress_to_year(congress, year = "first")
}
\arguments{
\item{congress}{Congress number (integer). Accepts a single value or a vector.}

\item{year}{Which year of the term to return: \code{"first"} (default) or \code{"second"}.}
}
\value{
A numeric value or vector of calendar years.
}
\description{
Convert a Congress number to the calendar year(s) of that session.
Returns either the first or second year of the two-year term.
}
\examples{
congress_to_year(116)
# 2019
congress_to_year(116, year = "second")
# 2020
congress_to_year(c(100, 110, 116))
# 1987 2007 2019
}
