% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_to_congress.R
\name{year_to_congress}
\alias{year_to_congress}
\title{Year to Congress}
\usage{
year_to_congress(year)
}
\arguments{
\item{year}{A numeric year or vector of years (e.g., \code{2019}).}
}
\value{
An integer value or vector of Congress numbers.
}
\description{
Convert a calendar year to the corresponding Congress number.
}
\examples{
year_to_congress(2019)
# 116
year_to_congress(c(1987, 2007, 2019))
# 100 110 116
}
