% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_to_year.R
\name{congress_to_year}
\alias{congress_to_year}
\title{Congress to Year}
\usage{
congress_to_year(congress, year = "first")
}
\arguments{
\item{congress}{the numerical signifier for congress. Either a single value of a vector of values.}

\item{year}{Which year of the Congress to return. Rakes either 'first' or 'second', defaults to "first"}
}
\value{
a numeric value or vector of values of the years which correspond the Congress
}
\description{
This function takes the number of a Congress and returns the years that Congress was in session. 
It will return either the first or second year of the session of Congress.
}
