% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_to_year.R

\name{stdz_district}
\alias{stdz_district}
\title{Standardize District}
\usage{
stdz_district <- function(full_district = character(), state, 
district, denote_sen = "00", denote_al = "01", sep = "-", old_sep = NA, sep_loc = NA)
}
\arguments{
\item{full_district}{district or vector of districts.}

\item{state}{state abbreviation or vector of state abbreviations.}

\item{district}{district number or vector of district numbers.}

\item{denote_al}{character to denote at large districts, in the district column, defaults to 01.}

\item{sep}{separator to use in the new standardized district names, if old_sep is not specified it is also the separator input districts are split on.}

\item{sep_old}{separator to split input districts on. Overrides sep for this role.}

\item{sep_loc}{location to split state from district. Use if there is no separator between state abbreviation and district number in the input.}

}
\value{
A vector of standardized districts
}
\description{
This function takes a district, vector of districts or state and district seperately and returns standardized district names.
}
